idproof/requirements.txt
Pillow==10.0.0
Django==5.2.3
django-simple-captcha==0.6.0
